Rewrite parser, again
Rewrite the parser with an actual grammar this time.  It’s a bit ugly in
places – use of the 'manyTill'' function is especially nasty and ought
to be replaced with more <|> combinators, as in the 'literateFile'
definition.

Many thanks to Chelsea Voss <csvoss@mit.edu> for helping me figure out
the grammar for this.